Modify POST request for user signup
