+++ package.json
@@ -4,12 +4,13 @@
   "description": "Auto Mart is an online marketplace for automobiles of diverse makes, model or body type where users can sell their cars or buy from trusted dealerships or private sellers. ",
   "main": "src/index.js",
   "scripts": {
-    "test": "npm run lint && NODE_ENV=test mocha",
+    "test": "nyc mocha --recursive",
     "clean:build": "rm -R dist",
     "build": "babel src --out-dir dist",
     "start": "node dist/index.js",
     "start:dev": "nodemon --exec babel-node src/index.js ",
-    "watch": "nodemon --exec babel-node --presets babel-preset-env src/index.js"
+    "watch": "nodemon --exec babel-node --presets babel-preset-env src/index.js",
+    "coverage": "nyc report --reporter=lcov > coverage.lcov && codecov"
   },
   "repository": {
     "type": "git",
@@ -49,6 +50,7 @@
     "eslint-plugin-import": "^2.17.2",
     "eslint-plugin-react": "^7.11.0",
     "mocha": "^6.1.4",
-    "nodemon": "^1.19.0"
+    "nodemon": "^1.19.0",
+    "nyc": "^14.1.1"
   }
 }
