--- src/database/seeders/user-seed.js
@@ -1,3 +1,0 @@
-import User from '../../../src/database/models/user';
-
-class User
